+++ CHANGELOG.md
@@ -46,7 +46,7 @@
 * New command `projectile-run-term` (<kbd>C-c p x t</kbd>).
 * Let user unignore files in `.projectile` with the ! prefix.
 * Add a command to add all projects in a directory to the cache (`projectile-discover-projects-in-directory`).
-* Add a command to list dirty projects (for version controlled projects)
+* Add a command to list dirty version controlled projects (`projectile-browse-dirty-projects`).
 
 ### Changes